move trpcroutes so that we can handle the auth routes one in a separate file
