+++ tests/integration/v2/TrpcRoutes.test.ts
@@ -1,0 +1,98 @@
+import { Server } from "http";
+import request from "supertest";
+import logger from "../../../src/bootstrap/logger";
+import startServer from "../../../src/bootstrap/app";
+import { clearPrismaDb } from "../helpers";
+import initializeDb, { ExtendedPrismaClient } from "../../../src/bootstrap/prisma-db";
+import UserDAO from "../../../src/DAO/v2/UserDAO";
+import { handleExitInTest, registerCleanupCallback } from "../../../src/bootstrap/shutdownHandler";
+
+let app: Server;
+let userDAO: UserDAO;
+let prisma: ExtendedPrismaClient;
+
+async function shutdown() {
+    try {
+        await handleExitInTest();
+    } catch (err) {
+        logger.error(`Error while shutting down: ${err}`);
+    }
+}
+
+beforeAll(async () => {
+    logger.debug("~~~~~~TRPC ROUTES BEFORE ALL~~~~~~");
+    app = await startServer();
+    // Initialize Prisma first, independently of server startup
+    prisma = initializeDb(process.env.DB_LOGS === "true");
+    registerCleanupCallback(async () => {
+        await prisma.$disconnect();
+    });
+    userDAO = new UserDAO(prisma.user);
+    return app;
+});
+
+afterAll(async () => {
+    logger.debug("~~~~~~TRPC ROUTES AFTER ALL~~~~~~");
+    const shutdownResult = await shutdown();
+    if (app) {
+        app.close(() => {
+            logger.debug("CLOSED SERVER");
+        });
+    }
+    return shutdownResult;
+});
+
+describe("tRPC API endpoints", () => {
+    afterEach(async () => {
+        return await clearPrismaDb(prisma);
+    });
+
+    describe("tRPC Error Handling", () => {
+        it("should return properly formatted tRPC error for malformed JSON", async () => {
+            await request(app)
+                .post("/v2/auth.login.sendResetEmail")
+                .set("Content-Type", "application/json")
+                .send("invalid json")
+                .expect(400);
+
+            // JSON handling is handled by the routing-controllers ErrorHandler middleware for now,
+            // but leaving this here for future reference:
+
+            // expect(body.error).toMatchObject({
+            //     code: -32700, // tRPC PARSE_ERROR code
+            //     message: "Invalid JSON in request body",
+            // });
+        });
+
+        it("should handle HTTP method not allowed gracefully", async () => {
+            await request(app).get("/v2/auth.login.sendResetEmail").expect(404); // tRPC route doesn't exist for GET method
+        });
+    });
+
+    describe("tRPC Middleware Integration", () => {
+        it("should integrate with Express middleware stack", async () => {
+            // Test that CORS, compression, etc. work with tRPC routes
+            const response = await request(app)
+                .options("/v2/auth.login.sendResetEmail")
+                .set("Origin", "http://localhost:3030") // Set a valid origin for CORS
+                .expect(204);
+
+            // Should have CORS headers that are actually present in the response
+            expect(response.headers).toHaveProperty("access-control-allow-credentials");
+            expect(response.headers).toHaveProperty("access-control-allow-methods");
+        });
+
+        it("should handle request timeout properly", async () => {
+            // This is a basic test - in practice you'd want to test with a slow endpoint
+            const testUser = { email: "test@example.com", password: "testpassword123" };
+            await userDAO.createUsers([{ email: testUser.email, password: testUser.password }]);
+
+            const start = Date.now();
+            await request(app).post("/v2/auth.login.sendResetEmail").send({ email: testUser.email }).expect(200);
+            const duration = Date.now() - start;
+
+            // Should complete within reasonable time (not timeout)
+            expect(duration).toBeLessThan(5000);
+        });
+    });
+});
